% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/append_time_dependent_analysis_tables_to_sqlite.R
\name{append_time_dependent_analysis_tables_to_sqlite}
\alias{append_time_dependent_analysis_tables_to_sqlite}
\title{append_time_dependent_analysis_tables_to_sqlite}
\usage{
append_time_dependent_analysis_tables_to_sqlite(
  exportFolder,
  sqliteDbPath,
  cohort_id = ""
)
}
\arguments{
\item{sqliteDbPath}{}
}
\value{
vector
}
\description{
append_time_dependent_analysis_tables_to_sqlite
}
