% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresql_upload.R
\name{combination_exists}
\alias{combination_exists}
\title{Function to check if the combination of cohort ID and database ID exists}
\usage{
combination_exists(schema_name, cohort_table, cohort_id, database_id)
}
\arguments{
\item{database_id}{}
}
\description{
Function to check if the combination of cohort ID and database ID exists
}
