% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_concepts_correlation.R
\name{conceptsCorrelationServer}
\alias{conceptsCorrelationServer}
\title{conceptsCorrelationServer}
\usage{
conceptsCorrelationServer(id, list_data)
}
\arguments{
\item{list_data}{}
}
\description{
conceptsCorrelationServer
}
