% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresql_upload.R
\name{create_table}
\alias{create_table}
\title{Function to create a table based on the data frame structure}
\usage{
create_table(table_name, data)
}
\arguments{
\item{data}{}
}
\description{
Function to create a table based on the data frame structure
}
