% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_Interface_Functions.R
\name{download_csvs}
\alias{download_csvs}
\title{Function to retrieve data based on cohort ID and database ID ---------------}
\usage{
download_csvs(schema_name, cohort_id, database_id, folder_path)
}
\arguments{
\item{folder_path}{}
}
\description{
Function to retrieve data based on cohort ID and database ID ---------------
}
