% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cohort_against_databases.R
\name{evaluate_cohort_against_databases}
\alias{evaluate_cohort_against_databases}
\title{evaluate_cohort_against_databases}
\usage{
evaluate_cohort_against_databases(
  CohortID,
  databases,
  session_user,
  atlas_password,
  atlas_user_name = session_user,
  databases_csv = "PL_App_Databases.csv",
  upload_to_s3 = FALSE,
  sink_log = FALSE,
  force_evaluation = FALSE
)
}
\arguments{
\item{databases_csv}{}
}
\description{
evaluate_cohort_against_databases
}
