% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql_conection.R
\name{execute_command}
\alias{execute_command}
\title{execute_command}
\usage{
execute_command(query_sql)
}
\arguments{
\item{query_sql}{}
}
\description{
Connect to redshift and only execute an insert without expecting a result set
}
