% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresql_upload.R
\name{filter_data_columns}
\alias{filter_data_columns}
\title{filter_data_columns}
\usage{
filter_data_columns(table_name, data)
}
\arguments{
\item{data}{}
}
\description{
Function to filter data frame columns to match the database table columns
}
