% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresql_upload.R
\name{generate_insert_query}
\alias{generate_insert_query}
\title{Function to generate an SQL INSERT statement for a data frame}
\usage{
generate_insert_query(table_name, data)
}
\arguments{
\item{data}{}
}
\description{
Function to generate an SQL INSERT statement for a data frame
}
