% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sqlite_filepath.R
\name{generate_sqlite_filepath}
\alias{generate_sqlite_filepath}
\title{generate_sqlite_filepath}
\usage{
generate_sqlite_filepath(
  scratch_space_name,
  username,
  password,
  atlas_user_name,
  atlas_password,
  CohortID,
  server,
  multiple_databaseCodes,
  selectedDatabase,
  export_dir = file.path(getwd(), "export"),
  drop_cohort_stats_tables = FALSE,
  drivers_dir = paste(getwd(), "drivers", sep = "/"),
  send_notification_email = FALSE,
  session_user = NULL,
  schema_name = "phenotype_library"
)
}
\arguments{
\item{selectedDatabase}{}
}
\description{
generate_sqlite_filepath
}
