% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_Interface_Functions.R
\name{generate_sqlite_filepath_from_database}
\alias{generate_sqlite_filepath_from_database}
\title{generate_sqlite_filepath_from_database}
\usage{
generate_sqlite_filepath_from_database(
  scratch_space_name,
  username,
  password,
  atlas_user_name,
  atlas_password,
  CohortID,
  server,
  schema_name,
  multiple_databaseCodes,
  export_dir
)
}
\arguments{
\item{multiple_databaseCodes}{}
}
\description{
generate_sqlite_filepath_from_database
}
