% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cdm_schema}
\alias{get_cdm_schema}
\title{get_cdm_schema}
\usage{
get_cdm_schema(selectedDatabase, databases_csv = "PL_App_Databases.csv")
}
\arguments{
\item{selectedDatabase}{}
}
\value{
CDM schema name
}
\description{
get_cdm_schema
}
