% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Clinical_Code_List_Function.R
\name{get_clinical_code_list}
\alias{get_clinical_code_list}
\title{get_clinical_code_list}
\usage{
get_clinical_code_list(
  atlas_user_name = NULL,
  atlas_password = NULL,
  cohortTable = NULL,
  CohortID = NULL
)
}
\arguments{
\item{CohortID}{}
}
\description{
get_clinical_code_list
}
