% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cohort_table_for_evaluation}
\alias{get_cohort_table_for_evaluation}
\title{get_cohort_table_for_evaluation}
\usage{
get_cohort_table_for_evaluation()
}
\value{
table name used in Cohort evaluation
}
\description{
get_cohort_table_for_evaluation
}
