% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_codes.R
\name{get_database_codes}
\alias{get_database_codes}
\title{get_database_codes}
\usage{
get_database_codes(selected_databases, PL_app_dbs)
}
\arguments{
\item{PL_app_dbs}{}
}
\value{
list
}
\description{
get_database_codes
}
