% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql_conection.R
\name{get_postgresql_RO_connection_details}
\alias{get_postgresql_RO_connection_details}
\title{get_postgresql_RO_connection_details}
\usage{
get_postgresql_RO_connection_details()
}
\description{
Set the READ-ONLY connection details
}
