% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresql_upload.R
\name{get_table_schema}
\alias{get_table_schema}
\title{get_table_schema}
\usage{
get_table_schema(table_name)
}
\arguments{
\item{table_name}{}
}
\description{
Create query to retrieve existing table (Function to retrieve the existing table schema using run_query_RW from global.R)
}
