% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_evaluation.R
\name{launch_cohort_evaluation}
\alias{launch_cohort_evaluation}
\title{launch_cohort_evaluation}
\usage{
launch_cohort_evaluation(
  datatable_rows_selected,
  dataBase,
  CohortID,
  atlas_user_name,
  atlas_password,
  home_page_df,
  phenotype_details,
  PL_app_dbs,
  S3_bucket_server,
  S3_bucket_cdsqlite_path,
  S3_bucket_region,
  run_Environment,
  redshift_username,
  redshift_password,
  session_user,
  session = NULL,
  unlink_outputs = FALSE,
  export_dir = file.path(getwd(), "export"),
  force_evaluation = FALSE,
  drivers_dir = paste(getwd(), "drivers", sep = "/"),
  postgresql_schema_name = "phenotype_library"
)
}
\arguments{
\item{force_evaluation}{}
}
\description{
launch_cohort_evaluation
}
