% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_evaluate_cohort.R
\name{mod_evaluate_cohort_ui}
\alias{mod_evaluate_cohort_ui}
\alias{mod_evaluate_cohort_server}
\title{mod_evaluate_cohort_ui}
\usage{
mod_evaluate_cohort_ui(id)

mod_evaluate_cohort_server(
  id,
  databases_csv,
  session_user,
  atlas_password,
  atlas_user_name,
  upload_to_s3 = FALSE
)
}
\arguments{
\item{id}{}

\item{databases_csv}{}
}
\description{
mod_evaluate_cohort_ui

mod_evaluate_cohort_server
}
