% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_codes.R
\name{parse_multiple_databaseCodes}
\alias{parse_multiple_databaseCodes}
\title{parse_multiple_databaseCodes}
\usage{
parse_multiple_databaseCodes(
  multiple_databaseCodes,
  databases_csv = "PL_App_Databases.csv"
)
}
\arguments{
\item{databases_csv}{}
}
\value{
vector of databases
}
\description{
parse_multiple_databaseCodes
}
