% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prevalence_changes.R
\name{prevalenceChangesServer}
\alias{prevalenceChangesServer}
\title{prevalenceChangesServer}
\usage{
prevalenceChangesServer(
  id,
  cohort_id,
  selected_dbs_df,
  sqlite_db_path,
  path_to_driver = "drivers"
)
}
\arguments{
\item{list_data}{}
}
\description{
prevalenceChangesServer
}
