% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_Interface_Functions.R
\name{process_cohort_data}
\alias{process_cohort_data}
\title{process_cohort_data}
\usage{
process_cohort_data(schema_name, CohortID, multiple_databaseCodes, export_dir)
}
\arguments{
\item{multiple_databaseCodes}{}
}
\description{
Main function to handle cohort data from database
}
