% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_mail_sqlite_to_s3}
\alias{send_mail_sqlite_to_s3}
\title{send_mail_sqlite_to_s3}
\usage{
send_mail_sqlite_to_s3(uploaded_s3_object, CohortID, databases, session_user)
}
\arguments{
\item{session_user}{}
}
\value{
bool
}
\description{
send_mail_sqlite_to_s3
}
