% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_mail_zip_to_database}
\alias{send_mail_zip_to_database}
\title{send_mail_zip_to_database}
\usage{
send_mail_zip_to_database(CohortID, databases, session_user)
}
\arguments{
\item{session_user}{}
}
\description{
send_mail_zip_to_database
}
