% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_Interface_Functions.R
\name{sqlite_path}
\alias{sqlite_path}
\title{sqlite_path}
\usage{
sqlite_path(
  scratch_space_name = NULL,
  username = NULL,
  password = NULL,
  atlas_user_name = NULL,
  atlas_password = NULL,
  CohortID = NULL,
  server = NULL,
  selectedDatabase = NULL,
  unlink_outputs = TRUE,
  export_dir = file.path(getwd(), "export"),
  force_evaluation = FALSE
)
}
\arguments{
\item{selectedDatabase}{}
}
\description{
This function returns arguments that are required to run the server part of Cohort Diagnostics App
}
