% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_Interface_Functions.R
\name{sqlite_path_multiple}
\alias{sqlite_path_multiple}
\title{sqlite_path_multiple}
\usage{
sqlite_path_multiple(
  scratch_space_name = NULL,
  username = NULL,
  password = NULL,
  atlas_user_name = NULL,
  atlas_password = NULL,
  CohortID = NULL,
  server = NULL,
  selectedDatabase = NULL,
  multiple_databaseCodes = NULL,
  total_db_count = 0,
  db_count = 0,
  ExistingCohortBit = NULL,
  IsMultipleDB_Flag = NULL,
  unlink_outputs = TRUE,
  export_dir = file.path(getwd(), "export"),
  drivers_dir = paste(getwd(), "drivers", sep = "/"),
  force_evaluation = FALSE,
  schema_name = "phenotype_library",
  send_notification_email = FALSE,
  session_user = NULL
)
}
\arguments{
\item{IsMultipleDB_Flag}{}
}
\description{
Function to generate sqlite path for selected multiple databases retrieving from postfresql database
}
