% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_time_dependent.R
\name{timeDependentServer}
\alias{timeDependentServer}
\title{timeDependentServer}
\usage{
timeDependentServer(
  id,
  cohort_id,
  selected_dbs_df,
  sqlite_db_path,
  path_to_driver = "drivers",
  threshold = 30L,
  n_plots_in_row = 4L,
  small_plot_with = 3L
)
}
\arguments{
\item{list_data}{}
}
\description{
timeDependentServer
}
