% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\title{Show notification with \code{shinyWidgets::show_toast}}
\usage{
toast(..., timer = 10 * 1000L)
}
\arguments{
\item{...}{arguments for \code{shinyWidgets::show_toast}}

\item{timer}{}
}
\description{
Wrapper of \code{shinyWidgets::show_toast}
}
