% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postresql_upload.R
\name{upload_csvs_from_zip_to_db}
\alias{upload_csvs_from_zip_to_db}
\title{Function to upload multiple CSV files from a zip archive to existing database tables}
\usage{
upload_csvs_from_zip_to_db(
  zip_file_path,
  CohortID,
  selectedDatabase,
  overwrite = FALSE,
  append = TRUE,
  tablePrefix = "phenotype_library"
)
}
\arguments{
\item{append}{}
}
\description{
Function to upload multiple CSV files from a zip archive to existing database tables
}
