% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cohort_against_databases.R
\name{upload_sqlite_to_s3}
\alias{upload_sqlite_to_s3}
\title{upload_sqlite_to_s3}
\usage{
upload_sqlite_to_s3(
  sqlite_path,
  s3_bucket_name = Sys.getenv("S3_bucket_server"),
  s3_key_prefix = "CD_SQLITE/OPTUM_EHR"
)
}
\arguments{
\item{s3_key_prefix}{}
}
\description{
upload_sqlite_to_s3
}
